bot_python_mysql/requirements.txt
mysql-connector-python>=8.0
python-dotenv>=1.0
python-telegram-bot>=20.0
